# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Patient Emergency & Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Priority = typeof Priority[keyof typeof Priority];


export const Priority = {
  RED: 'RED',
  YELLOW: 'YELLOW',
  GREEN: 'GREEN',
} as const;

export type PatientStatus = typeof PatientStatus[keyof typeof PatientStatus];


export const PatientStatus = {
  active: 'active',
  handled: 'handled',
} as const;

export interface Patient {
  id: number;
  name: string;
  room_no: string;
  bed_no: string;
  symptoms: string;
  priority: Priority;
  status: PatientStatus;
  created_at: string;
}

export interface AdmitPatientRequest {
  name: string;
  room_no: string;
  bed_no?: string;
  symptoms: string;
}

export interface UpdatePatientRequest {
  name?: string;
  room_no?: string;
  bed_no?: string;
  symptoms?: string;
  status?: PatientStatus;
  priority?: Priority;
}

export interface PatientStats {
  total: number;
  active: number;
  handled: number;
  red: number;
  yellow: number;
  green: number;
}

export interface DeleteResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}