# lib/api-zod/src/generated/types/admitPatientRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Patient Emergency & Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface AdmitPatientRequest {
  name: string;
  room_no: string;
  bed_no?: string;
  symptoms: string;
}