# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Patient Emergency & Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from './patientStatus';
import type { Priority } from './priority';

export interface Patient {
  id: number;
  name: string;
  room_no: string;
  bed_no: string;
  symptoms: string;
  priority: Priority;
  status: PatientStatus;
  created_at: Date;
}