# lib/api-zod/src/generated/types/patientStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Patient Emergency & Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export type PatientStatus = typeof PatientStatus[keyof typeof PatientStatus];


export const PatientStatus = {
  active: 'active',
  handled: 'handled',
} as const;