# lib/api-zod/src/generated/types/priority.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Patient Emergency & Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export type Priority = typeof Priority[keyof typeof Priority];


export const Priority = {
  RED: 'RED',
  YELLOW: 'YELLOW',
  GREEN: 'GREEN',
} as const;